+++ components/wasm-opt-cxx-sys/build.rs
@@ -1,13 +1,15 @@
 fn main() -> anyhow::Result<()> {
+    println!("cargo:rerun-if-changed=src/shims.h");
+
     let mut builder = cxx_build::bridge("src/lib.rs");
 
     {
         let target_env = std::env::var("CARGO_CFG_TARGET_ENV")?;
 
         let flags: &[_] = if target_env != "msvc" {
-            &["-std=c++17", "-Wno-unused-parameter", "-DBUILD_LLVM_DWARF"]
+            &["-std=c++20", "-Wno-unused-parameter", "-DBUILD_LLVM_DWARF"]
         } else {
-            &["/std:c++17", "/DBUILD_LLVM_DWARF"]
+            &["/std:c++20", "/DBUILD_LLVM_DWARF"]
         };
 
         for flag in flags {